requirements.txt
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
greenlet==3.2.3
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
